Improve file organization
