+++ app/src/main/java/com/example/studyapp/ui/screens/subtopics/PaneContent.kt
@@ -72,7 +72,6 @@ fun SubtopicsPaneContent(
     }
 }
 
-
 @Composable
 private fun SubtopicsLazyColumn(
     filteredSubtopics: List<Subtopic>,
