+++ app/src/main/java/com/example/studyapp/ui/screens/subtopics/dialogs/EditTopicDialog.kt
@@ -1,0 +1,57 @@
+package com.example.studyapp.ui.screens.subtopics.dialogs
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.example.studyapp.R
+import com.example.studyapp.data.Topic
+
+@Composable
+fun EditTopicDialog(
+    topic: Topic?,
+    modifier: Modifier = Modifier,
+    onDismiss: () -> Unit,
+    onSave: (Topic) -> Unit
+) {
+    var topicTitle by rememberSaveable { mutableStateOf(topic?.title ?: "") }
+    AlertDialog(
+        onDismissRequest = { onDismiss() },
+        title = { Text(stringResource(R.string.edit_topic)) },
+        text = {
+            OutlinedTextField(
+                value = topicTitle,
+                onValueChange = { topicTitle = it },
+                label = { Text(stringResource(R.string.title)) })
+        },
+        confirmButton = {
+            TextButton(
+                onClick = {
+                    if (topic != null) {
+                        onSave(topic.copy(title = topicTitle))
+                    } else {
+                        onSave(Topic(title = topicTitle, checked = false))
+                    }
+                    onDismiss()
+                }) {
+                Text(stringResource(R.string.save))
+            }
+        },
+        dismissButton = {
+            TextButton(
+                onClick = { onDismiss() }) {
+                Text(stringResource(R.string.cancel))
+            }
+        },
+        modifier = modifier
+    )
+}
+
+
